Role dodate

+++ app/src/main/java/com/example/dimitrije/pmsu/adapters/CommentAdapter.java
@@ -19,6 +19,7 @@ import com.example.dimitrije.pmsu.R;
 import com.example.dimitrije.pmsu.model.Comment;
 import com.example.dimitrije.pmsu.service.CommentService;
 import com.example.dimitrije.pmsu.service.ServiceUtils;
+import com.google.gson.Gson;
 
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
@@ -55,6 +56,9 @@ public class CommentAdapter extends ArrayAdapter<Comment> {
         if(view == null){
             view = LayoutInflater.from(getContext()).inflate(R.layout.comment_list_layout,viewGroup,false);
         }
+        sharedPreferences = getContext().getSharedPreferences(LoginActivity.MyPres, Context.MODE_PRIVATE);
+
+        final String userPref = sharedPreferences.getString(LoginActivity.Username, "");
 
         TextView title_view = view.findViewById(R.id.titleComentList);
         TextView author_view = view.findViewById(R.id.authorCommentList);
@@ -78,12 +82,15 @@ public class CommentAdapter extends ArrayAdapter<Comment> {
 
         commentService = ServiceUtils.commentService;
 
-        sharedPreferences = getContext().getSharedPreferences(MyPreferences, Context.MODE_PRIVATE);
         deleteButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                deleteComment(comment.getId());
-                Toast.makeText(getContext(), "Deleted", Toast.LENGTH_SHORT).show();
+                if(!userPref.equals(comment.getAuthor().getUsername()) || userPref.equals("")){
+                    Toast.makeText(getContext(), "You cant delete your comment", Toast.LENGTH_SHORT).show();
+                }else{
+                    deleteComment(comment.getId());
+                    Toast.makeText(getContext(), "Deleted", Toast.LENGTH_SHORT).show();
+                }
             }
         });
 
@@ -91,44 +98,53 @@ public class CommentAdapter extends ArrayAdapter<Comment> {
         likeButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                likeCount = comment.getLikes();
-                comment.setLikes(likeCount + 1);
 
-                Call<Comment> call = commentService.likeDislike(comment, comment.getId());
+                if(!userPref.equals(comment.getAuthor().getUsername()) || userPref.equals("")){
+                    likeCount = comment.getLikes();
+                    comment.setLikes(likeCount + 1);
+                    Call<Comment> call = commentService.likeDislike(comment, comment.getId());
 
-                call.enqueue(new Callback<Comment>() {
-                    @Override
-                    public void onResponse(Call<Comment> call, Response<Comment> response) {
-                        Toast.makeText(getContext(), "Liked", Toast.LENGTH_SHORT).show();
-                    }
+                    call.enqueue(new Callback<Comment>() {
+                        @Override
+                        public void onResponse(Call<Comment> call, Response<Comment> response) {
+                            Toast.makeText(getContext(), "Liked", Toast.LENGTH_SHORT).show();
+                        }
 
-                    @Override
-                    public void onFailure(Call<Comment> call, Throwable t) {
+                        @Override
+                        public void onFailure(Call<Comment> call, Throwable t) {
 
-                    }
-                });
+                        }
+                    });
+                }else{
+                    Toast.makeText(getContext(), "You cant like your comment", Toast.LENGTH_SHORT).show();
+                }
             }
         });
 
         dislikeButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                dislikeCount = comment.getDislikes();
-                comment.setDislikes(dislikeCount + 1);
 
-                Call<Comment> call = commentService.likeDislike(comment, comment.getId());
+                if (!userPref.equals(comment.getAuthor().getUsername()) || userPref.equals("")) {
+                    dislikeCount = comment.getDislikes();
+                    comment.setDislikes(dislikeCount + 1);
 
-                call.enqueue(new Callback<Comment>() {
-                    @Override
-                    public void onResponse(Call<Comment> call, Response<Comment> response) {
-                        Toast.makeText(getContext(), "Disliked", Toast.LENGTH_SHORT).show();
-                    }
+                    Call<Comment> call = commentService.likeDislike(comment, comment.getId());
 
-                    @Override
-                    public void onFailure(Call<Comment> call, Throwable t) {
+                    call.enqueue(new Callback<Comment>() {
+                        @Override
+                        public void onResponse(Call<Comment> call, Response<Comment> response) {
+                            Toast.makeText(getContext(), "Disliked", Toast.LENGTH_SHORT).show();
+                        }
 
-                    }
-                });
+                        @Override
+                        public void onFailure(Call<Comment> call, Throwable t) {
+
+                        }
+                    });
+                }else{
+                    Toast.makeText(getContext(), "You cant dislike your comment", Toast.LENGTH_SHORT).show();
+                }
             }
         });
 
@@ -137,10 +153,14 @@ public class CommentAdapter extends ArrayAdapter<Comment> {
             public void onClick(View view) {
                SharedPreferences.Editor editor = sharedPreferences.edit();
 
-               editor.putInt("Comment", comment.getId());
-               editor.commit();
-               getContext().startActivity(new Intent(getContext(), EditCommentActivity.class));
-               editor.clear();
+               if(!userPref.equals(comment.getAuthor().getUsername()) || userPref.equals("")){
+                   Toast.makeText(getContext(), "You dont have permision to edit this comment", Toast.LENGTH_SHORT).show();
+               }else{
+                   Intent intent = new Intent(getContext(), EditCommentActivity.class);
+                   intent.putExtra("commentId",comment.getId());
+                   getContext().startActivity(intent);
+
+               }
             }
         });
